requirements.txt
tiktoken
langchain
langchain_core
openai
fastapi
uvicorn
python-dotenv
APScheduler
python-jose
psycopg2
pymongo
pymongo[srv]
xhtml2pdf
python-jose